tedSpider/requirement.txt
aliyun-python-sdk-core-v3==2.8.9
aliyun-python-sdk-kms==2.5.0
beautifulsoup4==4.6.0
bitarray==0.8.3
bs4==0.0.1
certifi==2018.4.16
chardet==3.0.4
crcmod==1.7
future==0.16.0
greenlet==0.4.13
idna==2.7
lxml==4.2.2
oss2==2.5.0
peewee==3.5.0
Pillow==5.1.0
pycryptodome==3.6.3
pyexecjs==1.5.1
pymongo==3.7.0
PyMySQL==0.8.1
requests==2.11.1
six==1.11.0
urllib3==1.23
wechatsogou==4.2.2
Werkzeug==0.14.1
xlrd==1.1.0
